wireguard-go: use proper Go modules instead of git submodule
Instead of having to do the insane submodule hack like before,
wireguard-go has finally gotten with the times and uses a proper Go
module with source code pinning via hashes (updatable with go-get). This
enables you to use the latest wireguard-go improvements, as well as Go
1.12.

This commit is entirely untested, but it appears to compile.