wireguard-go: relicense to MIT
This project seems to statically link to OpenVPN, which is GPL-2.0.
IANAL, but I'm pretty sure Apache-2.0 and GPL-2.0 don't get along
especially well, and this might lead to problems.

Since I'm the sole copyright listed on this file, it's easy enough to
just relicense it to a super liberal license, MIT, that mostly gets
along with everything.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>