Fix wrongly specified arm64 MacOS target
We (and Rust) refer to a 64-bit ARM architecture as `aarch64`.
The official name which apple uses is `arm64`, while the rest of the
world use `aarch64` to refer to a 64-bit ARM architecture. Since there
is no way to get `uname` to spit out `aarch64`, a simple
lookup+re-mapping is done on our end if the host is apple silicon
(`arm64`).