openssl: Use GitHub "mirror"
git.openssl.org has been down for several days now which has been
preventing us on Solus from being able to build Mullvad.
They do however have a GitHub "mirror" which actually appears
to be the main development repository that is synced to git.openssl.org
instead of the other way around. Since both the mullvadvpn-app and
mullvadvpn-app-binaries repositories are also hosted on GitHub it
makes more sense to use GitHub for OpenSSL as doing so drops a
build-time dependency on an external host.

Also change the URLs of the other git submodules to end
in `.git` for consistency.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>